user-workspace/backend/requirements.txt
flask==3.0.2
google-cloud-aiplatform==1.38.1
google-cloud-secret-manager==2.18.0
earthengine-api==0.1.375
python-dotenv==1.0.1
flask-cors==4.0.0
google-generativeai==0.3.2
